Add files via upload
SQL Coding project completed by following instructions from Instructor Alyx of Analyst Builder.

Throughout the process, we searched if residents' life expectancy correlated with the developed or developing country they are living in or individual BMI, and the results are, that people with lower BMI and living in a developed country tend to have high life expectancy. oppositely, people with higher BMI and living in developing countries have less life expectancy.